// backend/src/utils/types.ts
import { USER_ROLES } from 'src/generated/prisma/enums';

export type JwtPayloadType = {
  sub: number;
  email: string;
  role: USER_ROLES;
  isVerified: boolean;
};

export interface FacebookUser {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
}

export interface PaginatedResult<T> {
  data: T[];
  metadata: {
    totalItems: number;
    itemCount: number;
    itemsPerPage: number;
    totalPages: number;
    currentPage: number;
    hasNextPage: boolean;
    hasPreviousPage: boolean;
    [key: string]: any;
  };
}

export interface SendMailData {
  userId: number;
  email: string;
  code: string;
}

export enum EMAIL_JOB_NAMES {
  SEND_VERIFICATION_EMAIL = 'send-verification-email',
  SEND_RESET_PASSWORD_EMAIL = 'send-reset-password-email',
}